Retrieve HWLOC API version from hwloc.h header
The FindLIBHWLOC.cmake module previously used the config.h header
to retrieve the HWLOC library version. This header was not always
present depending on how the library was distributed. To address
this issue, we now rely on the HWLOC API version retrieved from
hwloc.h, the main HWLOC library header.